Feat: ListStats, ListStatsByFilter, and AlterStats/ Refactor : search.js
Added options to list stats with pagination, list stats by filter with pagination, and change stats. The search function needed to be refactored, and the searchStats function was added to handle only the stats and return the correct SQL. Additionally, a separate pagination file was created to handle only the stats.